+++ lib/services/database/database.service.d.ts
@@ -1,8 +1,9 @@
 /// <reference types="mongoose" />
+import { MongoClient } from 'mongodb';
 import { ConfigService } from '../config/config.service';
 export declare class DatabaseService {
     private configService;
     constructor(configService: ConfigService);
-    connect(): Promise<import("mongodb").Db>;
+    connect(): Promise<MongoClient>;
     mongooseConnect(): Promise<typeof import("mongoose")>;
 }