+++ lib/templates/es5.d.ts
@@ -1,2 +1,2 @@
-declare const _default: "\nmodule.exports = {\n  async up (db) {\n    return ['Up']\n  },\n\n  async down (db) {\n    return ['Down']\n  }\n}\n";
+declare const _default: "\nmodule.exports = {\n  async up (client) {\n    return ['Up']\n  },\n\n  async down (client) {\n    return ['Down']\n  }\n}\n";
 export default _default;