+++ lib/templates/es6.d.ts
@@ -1,2 +1,2 @@
-declare const _default: "\nexport async function up(db) {\n  return ['Up'];\n}\nexport async function down(db) {\n  return ['Down'];\n}\n";
+declare const _default: "\nexport async function up(client) {\n  return ['Up'];\n}\nexport async function down(client) {\n  return ['Down'];\n}\n";
 export default _default;