+++ src/templates/es6.ts
@@ -1,8 +1,8 @@
 export default `
-export async function up(db) {
+export async function up(client) {
   return ['Up'];
 }
-export async function down(db) {
+export async function down(client) {
   return ['Down'];
 }
 `;