+++ src/templates/native.ts
@@ -1,18 +1,22 @@
 export default `
-export async function up (db) {
-  await db
+export async function up (client) {
+  await client
+    .db()
     .collection('albums')
     .updateOne({ artist: 'The Beatles' }, { $set: { blacklisted: true } })
-  await db
+  await client
+    .db()
     .collection('albums')
     .updateOne({ artist: 'The Doors' }, { $set: { stars: 5 } })
 },
 
-export async function down (db) {
-  await db
+export async function down (client) {
+  await client
+    .db()
     .collection('albums')
     .updateOne({ artist: 'The Doors' }, { $set: { stars: 0 } })
-  await db
+  await client
+    .db()
     .collection('albums')
     .updateOne({ artist: 'The Beatles' }, { $set: { blacklisted: false } })
 }