+++ src/templates/typescript.ts
@@ -1,23 +1,27 @@
 export default `
-import { Db } from 'mongodb';
+import { MongoClient } from 'mongodb';
 
-export async function up(db: Db) {
-  await db
+export async function up(client: MongoClient) {
+  await client
+    .db()
     .collection('albums')
     .updateOne({ artist: 'The Beatles' }, { $set: { blacklisted: true } });
 
-  await db
+  await client
+    .db()
     .collection('albums')
     .updateOne({ artist: 'The Doors' }, { $set: { stars: 5 } });
 }
-
-export async function down(db: Db) {
-  await db
+export async function down(client: MongoClient) {
+  await client
+    .db()
     .collection('albums')
     .updateOne({ artist: 'The Doors' }, { $set: { stars: 0 } });
 
-  await db
+  await client
+    .db()
     .collection('albums')
     .updateOne({ artist: 'The Beatles' }, { $set: { blacklisted: false } });
 }
+
 `;